move drop mode from ili2db library to gui_utils and make comments and clean up in gui_utils
