+++ QgisModelBaker/libili2db/globals.py
@@ -53,12 +53,6 @@ displayDbIliMode = {
 }
 
 
-class DropMode(Enum):
-    YES = 1
-    NO = 2
-    ASK = 3
-
-
 class DbActionType(Enum):
     """Defines constants for generate, import data, or export actions of QgisModelBaker."""
